Split tool artifact preview helpers
