+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -45,19 +45,32 @@ final class ParityGateTests: XCTestCase {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
         let toolCardSurfaceText = try Self.appSourceText(named: "QuillCodeToolCardSurface.swift")
         let toolArtifactSurfaceText = try Self.appSourceText(named: "QuillCodeToolArtifactSurface.swift")
+        let artifactValueClassifierText = try Self.appSourceText(named: "ToolArtifactValueClassifier.swift")
+        let artifactImagePreviewText = try Self.appSourceText(named: "ToolArtifactImagePreviewBuilder.swift")
+        let artifactDocumentPreviewText = try Self.appSourceText(named: "ToolArtifactDocumentPreviewBuilder.swift")
+        let artifactTextPreviewText = try Self.appSourceText(named: "ToolArtifactTextPreviewBuilder.swift")
         let transcriptBuilderText = try Self.appSourceText(named: "WorkspaceTranscriptSurfaceBuilder.swift")
 
         XCTAssertTrue(toolCardSurfaceText.contains("public struct ToolCardState"), "Tool card surface state should live in a focused surface file.")
         XCTAssertTrue(toolArtifactSurfaceText.contains("public struct ToolArtifactState"), "Tool artifact surface state should live in a focused artifact surface file.")
-        XCTAssertTrue(toolArtifactSurfaceText.contains("enum ToolArtifactPreviewBuilder"), "Tool artifact preview construction should live beside artifact state.")
+        XCTAssertTrue(toolArtifactSurfaceText.contains("ToolArtifactValueClassifier.kind"), "Tool artifact state should delegate value classification.")
+        XCTAssertTrue(toolArtifactSurfaceText.contains("ToolArtifactImagePreviewBuilder.imagePreview"), "Tool artifact state should delegate image previews.")
+        XCTAssertTrue(toolArtifactSurfaceText.contains("ToolArtifactDocumentPreviewBuilder.documentPreview"), "Tool artifact state should delegate document previews.")
+        XCTAssertTrue(artifactValueClassifierText.contains("enum ToolArtifactValueClassifier"), "Artifact value classification should have a focused owner.")
+        XCTAssertTrue(artifactImagePreviewText.contains("enum ToolArtifactImagePreviewBuilder"), "Image preview construction should have a focused owner.")
+        XCTAssertTrue(artifactDocumentPreviewText.contains("enum ToolArtifactDocumentPreviewBuilder"), "Document preview construction should have a focused owner.")
+        XCTAssertTrue(artifactTextPreviewText.contains("enum ToolArtifactTextPreviewBuilder"), "Artifact text-preview file reading should have a focused owner.")
         XCTAssertTrue(toolArtifactSurfaceText.contains("public struct ToolArtifactDocumentPreview"), "Document preview contracts should live beside artifact state.")
         XCTAssertTrue(toolArtifactSurfaceText.contains("public struct ToolArtifactImagePreview"), "Image preview contracts should live beside artifact state.")
-        XCTAssertTrue(transcriptBuilderText.contains("ToolArtifactPreviewBuilder.textPreview"), "Transcript projection should request artifact previews through the extracted builder.")
+        XCTAssertTrue(transcriptBuilderText.contains("ToolArtifactTextPreviewBuilder.textPreview"), "Transcript projection should request artifact text previews through the extracted builder.")
         XCTAssertFalse(modelText.contains("public struct ToolCardState"), "WorkspaceModel should not own tool card surface state.")
         XCTAssertFalse(modelText.contains("public enum ToolCardStatus"), "WorkspaceModel should not own tool card status.")
         XCTAssertFalse(modelText.contains("public struct ToolArtifactState"), "WorkspaceModel should not own tool artifact surface state.")
-        XCTAssertFalse(modelText.contains("ToolArtifactPreviewBuilder.textPreview"), "WorkspaceModel should not own artifact-preview requests.")
-        XCTAssertFalse(toolCardSurfaceText.contains("enum ToolArtifactPreviewBuilder"), "Tool-card state should not own artifact preview construction.")
+        XCTAssertFalse(modelText.contains("ToolArtifactTextPreviewBuilder.textPreview"), "WorkspaceModel should not own artifact-preview requests.")
+        XCTAssertFalse(toolArtifactSurfaceText.contains("private static func documentPreview"), "Tool artifact state should not own document-preview classification.")
+        XCTAssertFalse(toolArtifactSurfaceText.contains("private static func isImagePreview"), "Tool artifact state should not own image-preview classification.")
+        XCTAssertFalse(toolArtifactSurfaceText.contains("private static func localArtifactFileURL"), "Tool artifact state should not own text-preview file reading.")
+        XCTAssertFalse(toolCardSurfaceText.contains("ToolArtifactTextPreviewBuilder"), "Tool-card state should not own artifact preview construction.")
         XCTAssertFalse(toolCardSurfaceText.contains("public enum ToolArtifactDocumentKind"), "Tool-card state should not own artifact document metadata.")
     }
